+++ rsconnect/api.py
@@ -921,10 +921,9 @@ class RSConnectExecutor:
         Builds a dictionary containing the version of Posit Connect that is running
         and the versions of Python installed there.
 
-        :return: a three-entry dictionary.  The key 'connect' will refer to the version
+        :return: a two-entry dictionary.  The key 'connect' will refer to the version
         of Connect that was found.  The key `python` will refer to a sequence of version
-        strings for all the versions of Python that are installed.  The key `conda` will
-        refer to data about whether Connect is configured to support Conda environments.
+        strings for all the versions of Python that are installed.
         """
 
         def _to_sort_key(text):
@@ -934,16 +933,12 @@ class RSConnectExecutor:
         server_settings = self.server_settings
         python_settings = self.python_info
         python_versions = sorted([item["version"] for item in python_settings["installations"]], key=_to_sort_key)
-        conda_settings = {
-            "supported": python_settings["conda_enabled"] if "conda_enabled" in python_settings else False
-        }
         return {
             "connect": server_settings["version"],
             "python": {
                 "api_enabled": python_settings["api_enabled"] if "api_enabled" in python_settings else False,
                 "versions": python_versions,
             },
-            "conda": conda_settings,
         }
 
     def make_deployment_name(self, title, force_unique):
@@ -1121,14 +1116,9 @@ class PositClient(HTTPServer):
         return response
 
     def create_output(self, name: str, application_type: str, project_id=None, space_id=None, render_by=None):
-        data = {
-            "name": name,
-            "space": space_id,
-            "project": project_id,
-            "application_type": application_type
-        }
+        data = {"name": name, "space": space_id, "project": project_id, "application_type": application_type}
         if render_by:
-            data['render_by'] = render_by
+            data["render_by"] = render_by
         response = self.post("/v1/outputs/", body=data)
         self._server.handle_bad_response(response)
         return response
@@ -1342,9 +1332,7 @@ class CloudService:
         app_store_version: typing.Optional[int],
     ) -> PrepareDeployOutputResult:
 
-        application_type = "static" if app_mode in [
-            AppModes.STATIC,
-            AppModes.STATIC_QUARTO] else "connect"
+        application_type = "static" if app_mode in [AppModes.STATIC, AppModes.STATIC_QUARTO] else "connect"
         logger.debug(f"application_type: {application_type}")
 
         render_by = "server" if app_mode == AppModes.STATIC_QUARTO else None
@@ -1362,11 +1350,13 @@ class CloudService:
                 space_id = None
 
             # create the new output and associate it with the current Posit Cloud project and space
-            output = self._rstudio_client.create_output(name=app_name,
-                                                        application_type=application_type,
-                                                        project_id=project_id,
-                                                        space_id=space_id,
-                                                        render_by=render_by)
+            output = self._rstudio_client.create_output(
+                name=app_name,
+                application_type=application_type,
+                project_id=project_id,
+                space_id=space_id,
+                render_by=render_by,
+            )
             app_id_int = output["source_id"]
         else:
             # this is a redeployment of an existing output