+++ rsconnect/bundle.py
@@ -116,7 +116,6 @@ class Manifest:
                 },
             }
 
-
         if image or env_management_py is not None or env_management_r is not None:
             self.data["environment"] = {}
             if image:
@@ -459,8 +458,9 @@ def write_manifest(
     Returns the list of filenames written.
     """
     manifest_filename = "manifest.json"
-    manifest = make_source_manifest(AppModes.JUPYTER_NOTEBOOK, environment, nb_name, None,
-                                    image, env_management_py, env_management_r)
+    manifest = make_source_manifest(
+        AppModes.JUPYTER_NOTEBOOK, environment, nb_name, None, image, env_management_py, env_management_r
+    )
     if hide_all_input:
         if "jupyter" not in manifest:
             manifest["jupyter"] = {}
@@ -541,8 +541,9 @@ def make_notebook_source_bundle(
     base_dir = dirname(file)
     nb_name = basename(file)
 
-    manifest = make_source_manifest(AppModes.JUPYTER_NOTEBOOK, environment, nb_name, None,
-                                    image, env_management_py, env_management_r)
+    manifest = make_source_manifest(
+        AppModes.JUPYTER_NOTEBOOK, environment, nb_name, None, image, env_management_py, env_management_r
+    )
     if hide_all_input:
         if "jupyter" not in manifest:
             manifest["jupyter"] = {}
@@ -595,8 +596,15 @@ def make_quarto_source_bundle(
     Returns a file-like object containing the bundle tarball.
     """
     manifest, relevant_files = make_quarto_manifest(
-        file_or_directory, inspect, app_mode, environment, extra_files, excludes,
-        image, env_management_py, env_management_r,
+        file_or_directory,
+        inspect,
+        app_mode,
+        environment,
+        extra_files,
+        excludes,
+        image,
+        env_management_py,
+        env_management_r,
     )
     bundle_file = tempfile.TemporaryFile(prefix="rsc_bundle")
 
@@ -873,8 +881,9 @@ def make_api_manifest(
     excludes.extend(list_environment_dirs(directory))
 
     relevant_files = create_file_list(directory, extra_files, excludes)
-    manifest = make_source_manifest(app_mode, environment, entry_point, None,
-                                    image, env_management_py, env_management_r)
+    manifest = make_source_manifest(
+        app_mode, environment, entry_point, None, image, env_management_py, env_management_r
+    )
 
     manifest_add_buffer(manifest, environment.filename, environment.contents)
 
@@ -938,8 +947,14 @@ def create_html_manifest(
     excludes.extend(["manifest.json"])
     excludes.extend(list_environment_dirs(deploy_dir))
 
-    manifest = Manifest(app_mode=AppModes.STATIC, entrypoint=entrypoint, primary_html=entrypoint,
-                        image=image, env_management_py=env_management_py, env_management_r=env_management_r)
+    manifest = Manifest(
+        app_mode=AppModes.STATIC,
+        entrypoint=entrypoint,
+        primary_html=entrypoint,
+        image=image,
+        env_management_py=env_management_py,
+        env_management_r=env_management_r,
+    )
     manifest.deploy_dir = deploy_dir
 
     file_list = create_file_list(path, extra_files, excludes, use_abspath=True)
@@ -1196,8 +1211,15 @@ def make_api_bundle(
     :return: a file-like object containing the bundle tarball.
     """
     manifest, relevant_files = make_api_manifest(
-        directory, entry_point, app_mode, environment, extra_files, excludes,
-        image, env_management_py, env_management_r,
+        directory,
+        entry_point,
+        app_mode,
+        environment,
+        extra_files,
+        excludes,
+        image,
+        env_management_py,
+        env_management_r,
     )
     bundle_file = tempfile.TemporaryFile(prefix="rsc_bundle")
 
@@ -1550,7 +1572,6 @@ def which_python(python: typing.Optional[str] = None):
 def inspect_environment(
     python,  # type: str
     directory,  # type: str
-    conda_mode=False,  # type: bool
     force_generate=False,  # type: bool
     check_output=subprocess.check_output,  # type: typing.Callable
 ):
@@ -1561,8 +1582,6 @@ def inspect_environment(
     or containing an "error" field if an error occurred.
     """
     flags = []
-    if conda_mode:
-        flags.append("c")
     if force_generate:
         flags.append("f")
     args = [python, "-m", "rsconnect.environment"]
@@ -1576,14 +1595,13 @@ def inspect_environment(
     return MakeEnvironment(**json.loads(environment_json))  # type: ignore
 
 
-def get_python_env_info(file_name, python, conda_mode=False, force_generate=False):
+def get_python_env_info(file_name, python, force_generate=False):
     """
     Gathers the python and environment information relating to the specified file
     with an eye to deploy it.
 
     :param file_name: the primary file being deployed.
     :param python: the optional name of a Python executable.
-    :param conda_mode: inspect the environment assuming Conda
     :param force_generate: force generating "requirements.txt" or "environment.yml",
     even if it already exists.
     :return: information about the version of Python in use plus some environmental
@@ -1591,7 +1609,7 @@ def get_python_env_info(file_name, python, conda_mode=False, force_generate=Fals
     """
     python = which_python(python)
     logger.debug("Python: %s" % python)
-    environment = inspect_environment(python, dirname(file_name), conda_mode=conda_mode, force_generate=force_generate)
+    environment = inspect_environment(python, dirname(file_name), force_generate=force_generate)
     if environment.error:
         raise RSConnectException(environment.error)
     logger.debug("Python: %s" % python)
@@ -1638,8 +1656,15 @@ def create_notebook_manifest_and_environment_file(
     """
     if (
         not write_notebook_manifest_json(
-            entry_point_file, environment, app_mode, extra_files, hide_all_input, hide_tagged_input,
-            image, env_management_py, env_management_r,
+            entry_point_file,
+            environment,
+            app_mode,
+            extra_files,
+            hide_all_input,
+            hide_tagged_input,
+            image,
+            env_management_py,
+            env_management_r,
         )
         or force
     ):
@@ -1691,8 +1716,9 @@ def write_notebook_manifest_json(
         if app_mode == AppModes.UNKNOWN:
             raise RSConnectException('Could not determine the app mode from "%s"; please specify one.' % extension)
 
-    manifest_data = make_source_manifest(app_mode, environment, file_name, None,
-                                         image, env_management_py, env_management_r)
+    manifest_data = make_source_manifest(
+        app_mode, environment, file_name, None, image, env_management_py, env_management_r
+    )
     if hide_all_input or hide_tagged_input:
         if "jupyter" not in manifest_data:
             manifest_data["jupyter"] = dict()
@@ -1789,8 +1815,14 @@ def create_voila_manifest(
     if isfile(voila_json_path):
         extra_files.append(voila_json_path)
 
-    manifest = Manifest(app_mode=AppModes.JUPYTER_VOILA, environment=environment, entrypoint=entrypoint,
-                        image=image, env_management_py=env_management_py, env_management_r=env_management_r)
+    manifest = Manifest(
+        app_mode=AppModes.JUPYTER_VOILA,
+        environment=environment,
+        entrypoint=entrypoint,
+        image=image,
+        env_management_py=env_management_py,
+        env_management_r=env_management_r,
+    )
     manifest.deploy_dir = deploy_dir
     if entrypoint and isfile(entrypoint):
         validate_file_is_notebook(entrypoint)
@@ -1880,8 +1912,17 @@ def create_api_manifest_and_environment_file(
     :return:
     """
     if (
-        not write_api_manifest_json(directory, entry_point, environment, app_mode, extra_files, excludes,
-                                    image, env_management_py, env_management_r)
+        not write_api_manifest_json(
+            directory,
+            entry_point,
+            environment,
+            app_mode,
+            extra_files,
+            excludes,
+            image,
+            env_management_py,
+            env_management_r,
+        )
         or force
     ):
         write_environment_file(environment, directory)
@@ -1919,8 +1960,9 @@ def write_api_manifest_json(
     etc.) that goes along with the manifest exists.
     """
     extra_files = validate_extra_files(directory, extra_files)
-    manifest, _ = make_api_manifest(directory, entry_point, app_mode, environment, extra_files, excludes,
-                                    image, env_management_py, env_management_r)
+    manifest, _ = make_api_manifest(
+        directory, entry_point, app_mode, environment, extra_files, excludes, image, env_management_py, env_management_r
+    )
     manifest_path = join(directory, "manifest.json")
 
     write_manifest_json(manifest_path, manifest)
@@ -1997,8 +2039,9 @@ def write_quarto_manifest_json(
     """
 
     extra_files = validate_extra_files(directory, extra_files)
-    manifest, _ = make_quarto_manifest(directory, inspect, app_mode, environment, extra_files, excludes,
-                                       image, env_management_py, env_management_r)
+    manifest, _ = make_quarto_manifest(
+        directory, inspect, app_mode, environment, extra_files, excludes, image, env_management_py, env_management_r
+    )
     manifest_path = join(directory, "manifest.json")
 
     write_manifest_json(manifest_path, manifest)
@@ -2017,7 +2060,6 @@ def create_python_environment(
     directory: str = None,
     force_generate: bool = False,
     python: str = None,
-    conda: bool = False,
 ):
     module_file = fake_module_file_from_directory(directory)
 
@@ -2028,7 +2070,7 @@ def create_python_environment(
     _warn_if_environment_directory(directory)
 
     # with cli_feedback("Inspecting Python environment"):
-    _, environment = get_python_env_info(module_file, python, conda, force_generate)
+    _, environment = get_python_env_info(module_file, python, force_generate)
 
     if force_generate:
         _warn_on_ignored_requirements(directory, environment.filename)